requirements.txt
Django==1.9.6
django-appconf==1.0.2
django-compressor==2.0
django-filter==0.13.0
djangorestframework==3.3.3
Markdown==2.6.6
mysqlclient==1.3.7
rcssmin==1.0.6
rjsmin==1.0.12
